[measureobjectintensity]: (3) Move code from library/module to library/funcitons/measurement
